Change default precision to 16 when saving Meshes and GridFunctions

+++ fem/gridfunc.hpp
@@ -673,7 +673,7 @@ public:
 
    /// Save the GridFunction to a file. The given @a precision will be used for
    /// ASCII output.
-   virtual void Save(const char *fname, int precision=8) const;
+   virtual void Save(const char *fname, int precision=16) const;
 
 #ifdef MFEM_USE_ADIOS2
    /// Save the GridFunction to a binary output stream using adios2 bp format.
